requirements.txt
langchain
langchain-community
langchainhub
openai
chromadb
bs4
cohere
tiktoken
fastapi
uvicorn
tortoise-orm
python-dotenv
asyncpg
